+++ src/HRBACHelper.php
@@ -31,7 +31,7 @@ class HRBACHelper
         } catch (UserHasNoBuiltInRolesException $e) {
             return null;
         } catch (PermissionNotFoundException $e) {
-            if (config('h-rbac.exceptIfPermissionNotFound', false)) {
+            if (config('h-rbac.riseExceptionIfPermissionNotFound', false)) {
                 throw $e;
             } else {
                 return null;